requirements.txt
argcomplete==1.9.3
awscli==1.16.88
boto3==1.9.78
botocore==1.12.78
certifi==2018.11.29
cfn-flip==1.1.0.post1
chardet==3.0.4
Click==7.0
colorama==0.3.9
Django==2.1.5
django-rest-framework==0.1.0
djangorestframework==3.9.0
docutils==0.14
durationpy==0.5
Flask==1.0.2
future==0.16.0
hjson==3.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
MarkupSafe==1.1.0
placebo==0.8.2
pyasn1==0.4.5
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
six==1.12.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.1
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.14.1
wsgi-request-logger==0.4.6
zappa==0.47.1
